PaymentTerm on Invoice Fix

+++ Entities/Invoice.cs
@@ -24,7 +24,7 @@ public class Invoice
 
     public decimal OffsetValue { get; set; }
 
-    public int PaymentTermID { get; set; }
+    public long PaymentTermID { get; set; }
 
     //Navigation Properties
